+++ src/vector.zig
@@ -1,21 +1,21 @@
 pub fn Vec(comptime size: u32) type {
-    return @Vector(size, f32);
+    return @Vector(size, f64);
 }
 
-pub fn filled(comptime size: usize, v: f32) Vec(size) {
+pub fn filled(comptime size: usize, v: f64) Vec(size) {
     return @splat(v);
 }
 
-pub fn dot(comptime size: usize, v1: Vec(size), v2: Vec(size)) f32 {
-    var acc: f32 = 0;
+pub fn dot(comptime size: usize, v1: Vec(size), v2: Vec(size)) f64 {
+    var acc: f64 = 0;
     inline for (0..size) |i| {
         acc += v1[i] * v2[i];
     }
     return acc;
 }
 
-pub fn len_squared(comptime size: usize, vec: Vec(size)) f32 {
-    var acc: f32 = 0;
+pub fn len_squared(comptime size: usize, vec: Vec(size)) f64 {
+    var acc: f64 = 0;
     inline for (0..size) |i| {
         acc += vec[i] * vec[i];
     }
